Fixing circular dependency
